Added Command Statuses
Using Rebus.Events to hook into Rebus event lifecycle hooks:

   - BeforeMessageSent -- Stamp new events with a unique
     SagaFlowCommandId, start date/time and initiating user. New events
     will be initialized with an attempt counter of 1.  Retried events
     due to errors will have their attempt counter incremented.

   - BeforeMessageHandled -- This will insert or update existing
     commands in the SagaFlowCommandStore.  For new commands a human
     readable version of their Parameters / names is recorded.  For all
     commands their progress is intialized to 0.

   - AfterMessageHandled -- Successfully completed commands are updated
     as successful, all ISagaFlowCommandSucceeededHandlers are invoked.
     For errored commands, the exception and stack trace are attached to
     command and saved to the SagaFlowCommandStore, all
     ISagaFlowCommandErroredHandlers are invoked.

Added IUsernameProvider, which is used by SagaFlow to resolve a username
to associate as the initiating user of the command.  A default
implementation is registered that simply looks up Name Identifier claim
of the current logged in principal.  A custom implementation can be
supplied .WithUsernameProvider<TUsernameProvider>() on the
SagaFlowOptionsBuilder.

Added ISagaFlowCommandStore, wich is used by SagaFlow to store all
commands invoked.  A default in-memory inmplementation is supplied,
which can be overriden with a customer implementation if required.

Added Rebus.SignalR package, which will add a SignalR hub for publishing
SagaFlow Updated, Succeeded and Errored events.  The SagaFlow.UI uses
the SignalR hub to update the web component sf-command-statuses.